# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Location Service",
        "title": "Location Service",
        "contact": {},
        "version": "1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/address/predict": {
            "get": {
                "description": "Get addresses by string.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get address suggestions by string",
                "operationId": "get-addresses-by-string",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.AutocompletePrediction"
                            }
                        }
                    },
                    "404": {
                        "description": "Please input a valid string",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/address/validate": {
            "get": {
                "description": "Get an address by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addresses"
                ],
                "summary": "Get address by ID",
                "operationId": "get-address-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Address"
                        }
                    },
                    "404": {
                        "description": "address not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Address": {
            "type": "object",
            "properties": {
                "addressComponents": {
                    "description": "The components of the address, such as street number, street name, city, state, etc."
                },
                "formattedAddress": {
                    "description": "The formatted address string.",
                    "type": "string"
                },
                "type": {
                    "description": "The type of the address, such as \"google\" or \"geoscape\".",
                    "type": "string"
                }
            }
        },
        "domain.AutocompletePrediction": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Formatted Address.",
                    "type": "string"
                },
                "place_id": {
                    "description": "address id.",
                    "type": "string"
                }
            }
        }
    }
}